refactor and setup settings
